+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - Bugfix: Fixed blocked user list sticking around when switching from a logged in user to being logged out. (#4437)
 - Bugfix: Fixed search popup ignoring setting for message scrollback limit. (#4496)
 - Bugfix: Fixed a memory leak that occurred when loading message history. This was mostly noticeable with unstable internet connections where reconnections were frequent or long-running instances of Chatterino. (#4499)
+- Bugfix: Fix visual glitches with smooth scrolling. (#4501)
 - Bugfix: Fixed Twitch channel-specific filters not being applied correctly. (#4529)
 - Bugfix: Fixed `/mods` displaying incorrectly when the channel has no mods. (#4546)
 - Bugfix: Fixed emote & badge tooltips not showing up when thumbnails were hidden. (#4509)